style:  reconstruct component files
